% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotopia_data_devices.R
\name{ecotopia_data_devices}
\alias{ecotopia_data_devices}
\title{Get User's Devices table from Ecotopia
Config should be set in config file.
File path is in the working directory and named as 'ecotopia.yml'.
ecotopia.yml example:
domain: domain_name
username: user_name
password: pass_word
token: token_string
token is optional
if token is provided, username and password will be ignored.}
\usage{
ecotopia_data_devices(max_devices = 1000, start_date = NULL)
}
\arguments{
\item{max_devices}{Limit max count of devices return. default is 1000.}

\item{start_date}{Require devices start from the specified time.}
}
\value{
A dataframe of devices information.
}
\description{
Get User's Devices table from Ecotopia
Config should be set in config file.
File path is in the working directory and named as 'ecotopia.yml'.
ecotopia.yml example:
domain: domain_name
username: user_name
password: pass_word
token: token_string
token is optional
if token is provided, username and password will be ignored.
}
\examples{
devices <- ecotopia_data_devices(
  max_devices = 1000,
  start_date = "2019-01-01"
)
}
