% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotopia_data_devices_acc.R
\name{ecotopia_data_devices_acc}
\alias{ecotopia_data_devices_acc}
\title{Get devices' ACC data from Ecotopia API}
\usage{
ecotopia_data_devices_acc(device_ids, show_progress = "OFF", start_time = NULL)
}
\arguments{
\item{device_ids}{Device IDs to get data from}

\item{show_progress}{Show download progress log
show_progress can be "OFF", "SIMPLE" or "FULL"}

\item{start_time}{Start time of the data to be downloaded}
}
\value{
List of devices contain list of vectors with their ACC data.
}
\description{
Get devices' ACC data from Ecotopia API
}
\examples{
ecotopia_data_devices_acc(
  c("device_id1", "device_id2", "device_id3"),
  show_progress = "SIMPLE"
)
}
