% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotopia_data_devices_env.R
\name{ecotopia_data_devices_env}
\alias{ecotopia_data_devices_env}
\title{Get devices' ENV data from Ecotopia API}
\usage{
ecotopia_data_devices_env(
  device_uuids,
  show_progress = "OFF",
  start_time = NULL
)
}
\arguments{
\item{device_uuids}{Device UUIDs to get data from}

\item{show_progress}{Show download progress log
show_progress can be "OFF", "SIMPLE" or "FULL"}

\item{start_time}{Start time of the data to be downloaded}
}
\value{
List of devices contain list of vectors with their ENV data.
}
\description{
Get devices' ENV data from Ecotopia API
}
\examples{
ecotopia_data_devices_env(
 c("device_id1", "device_id2", "device_id3"),
 show_progress = "SIMPLE"
)
}
