% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotopia_data_devices_odba.R
\name{ecotopia_data_devices_odba}
\alias{ecotopia_data_devices_odba}
\title{Get devices' ODBA data from Ecotopia API}
\usage{
ecotopia_data_devices_odba(
  device_ids,
  show_progress = "OFF",
  start_time = NULL
)
}
\arguments{
\item{device_ids}{Device IDs to get data from}

\item{show_progress}{Show download progress log
show_progress can be "OFF", "SIMPLE" or "FULL"}

\item{start_time}{Start time of the data to be downloaded}
}
\value{
List of devices contain list of vectors with their ODBA data.
}
\description{
Get devices' ODBA data from Ecotopia API
}
\examples{
ecotopia_data_devices_odba(
  c("device_id1", "device_id2", "device_id3"),
  show_progress = "SIMPLE"
)
}
