% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotopia_user_token.R
\name{ecotopia_user_token}
\alias{ecotopia_user_token}
\title{Get token by login to Ecotopia
Config should be set in config file.
File path is in the working directory and named as 'ecotopia.yml'.
ecotopia.yml example:
domain: domain_name
username: user_name
password: pass_word
token: token_string
token is optional
if token is provided, this function will return the token directly.}
\usage{
ecotopia_user_token()
}
\value{
A character string of the token for API calling.
}
\description{
Get token by login to Ecotopia
Config should be set in config file.
File path is in the working directory and named as 'ecotopia.yml'.
ecotopia.yml example:
domain: domain_name
username: user_name
password: pass_word
token: token_string
token is optional
if token is provided, this function will return the token directly.
}
\examples{
token <- ecotopia_user_token()
}
